Move EpdHw into hw mod
