+++ epd-waveshare-async/src/epd2in9.rs
@@ -12,7 +12,7 @@ use embedded_hal_async::delay::DelayNs;
 
 use crate::{
     buffer::{binary_buffer_length, split_low_and_high, BinaryBuffer},
-    comms::CommandDataSend as _,
+    hw::CommandDataSend as _,
     log::{debug, debug_assert},
     Epd, EpdHw,
 };
